Update inline documentation for public apis.
