+++ lib/src/charts/data/bars/multi_bar.dart
@@ -34,35 +34,36 @@ class MultiBar extends BarGroup {
   ///
   /// See Also: [BarGroup]
   ///
-  MultiBar(
-      {required super.xValue,
-      required this.yValues,
-      // defaults to series i.e. side by side
-      this.arrangement = BarGroupArrangement.series,
-      this.spacing = 10.0,
-      super.style,
-      super.padding})
-      : assert(yValues.isNotEmpty, "At least one yValue is required!"),
+  MultiBar({
+    required super.xValue,
+    required this.yValues,
+    // defaults to series i.e. side by side
+    this.arrangement = BarGroupArrangement.series,
+    this.spacing = 10.0,
+    super.style,
+    super.padding,
+  })  : assert(yValues.isNotEmpty, "At least one yValue is required!"),
         assert(spacing >= 0.0, "groupSpacing cannot be Negative!");
 
   /// Lerps between two [MultiBar]s for a factor [t]
   static MultiBar lerp(BarGroup? current, BarGroup target, double t) {
     if ((current is MultiBar?) && target is MultiBar) {
       return MultiBar(
-          xValue: lerpDouble(current?.xValue, target.xValue, t) as num,
-          yValues: BarData.lerpBarDataList(current?.yValues, target.yValues, t),
-          arrangement: target.arrangement,
-          spacing: lerpDouble(
-            current?.spacing,
-            target.spacing,
-            t,
-          ).asOrDefault(0.0),
-          style: BarDataStyle.lerp(
-            current?.style,
-            target.style,
-            t,
-          ),
-          padding: lerpDouble(current?.padding, target.padding, t) ?? 0.0);
+        xValue: lerpDouble(current?.xValue, target.xValue, t) as num,
+        yValues: BarData.lerpBarDataList(current?.yValues, target.yValues, t),
+        arrangement: target.arrangement,
+        spacing: lerpDouble(
+          current?.spacing,
+          target.spacing,
+          t,
+        ).asOrDefault(0.0),
+        style: BarDataStyle.lerp(
+          current?.style,
+          target.style,
+          t,
+        ),
+        padding: lerpDouble(current?.padding, target.padding, t) ?? 0.0,
+      );
     } else {
       throw Exception('Both current & target data should be of same series!');
     }
