+++ lib/src/charts/data/bars/simple_bar.dart
@@ -14,24 +14,26 @@ class SimpleBar extends BarGroup {
   /// Defines a Simple singular Bar with a Single Y-Value
   ///
   /// See Also: [BarGroup]
-  SimpleBar(
-      {required super.xValue,
-      required this.yValue,
-      super.style,
-      super.padding});
+  SimpleBar({
+    required super.xValue,
+    required this.yValue,
+    super.style,
+    super.padding,
+  });
 
   /// Lerps between two [SimpleBar]s for a factor [t]
   static SimpleBar lerp(BarGroup? current, BarGroup target, double t) {
     if ((current is SimpleBar?) && target is SimpleBar) {
       return SimpleBar(
-          xValue: lerpDouble(current?.xValue, target.xValue, t) as num,
-          yValue: BarData.lerp(current?.yValue, target.yValue, t),
-          style: BarDataStyle.lerp(
-            current?.style,
-            target.style,
-            t,
-          ),
-          padding: lerpDouble(current?.padding, target.padding, t) ?? 0.0);
+        xValue: lerpDouble(current?.xValue, target.xValue, t) as num,
+        yValue: BarData.lerp(current?.yValue, target.yValue, t),
+        style: BarDataStyle.lerp(
+          current?.style,
+          target.style,
+          t,
+        ),
+        padding: lerpDouble(current?.padding, target.padding, t) ?? 0.0,
+      );
     } else {
       throw Exception('Both current & target data should be of same series!');
     }
