+++ lib/src/charts/data/core/cartesian/cartesian_range.dart
@@ -56,7 +56,6 @@ class CartesianRangeResult with EquatableMixin {
   }
 
   @override
-  // TODO: implement props
   List<Object?> get props => [
         xUnitValue,
         yUnitValue,
